xzdualrecovery.sh: Modified the LED control function to match the one in chargemon/dualrecovery. Updated the fs unlocking mechanism, it should now kill a process only once and only if still running will attempt to kill it again.
chargemon.sh: updated LED target for the XSP
dualrecovery.sh: updated LED target for th XSP and updated the fs unlocking mechanism to match the one in xzdualrecovery.sh

+++ src/dualrecovery.sh
@@ -29,9 +29,10 @@ echo "START Dual Recovery at ${DATETIME}: STAGE 2." > ${LOG}
 #BOOTREC_CACHE="/dev/block/mmcblk0p25"
 BOOTREC_EXTERNAL_SDCARD_NODE="/dev/block/mmcblk1p1 b 179 32"
 BOOTREC_EXTERNAL_SDCARD="/dev/block/mmcblk1p1"
-REDLED=$(/sbin/busybox ls -1 /sys/class/leds|/sbin/busybox grep "red\|LED2_R")
-GREENLED=$(/sbin/busybox ls -1 /sys/class/leds|/sbin/busybox grep "green\|LED2_G")
-BLUELED=$(/sbin/busybox ls -1 /sys/class/leds|/sbin/busybox grep "blue\|LED2_B")
+
+REDLED=$(/sbin/busybox ls -1 /sys/class/leds|/sbin/busybox grep "red\|LED1_R")
+GREENLED=$(/sbin/busybox ls -1 /sys/class/leds|/sbin/busybox grep "green\|LED1_G")
+BLUELED=$(/sbin/busybox ls -1 /sys/class/leds|/sbin/busybox grep "blue\|LED1_B")
 
 # Defining functions
 ECHOL(){
@@ -256,10 +257,12 @@ if [ "$RECOVERYBOOT" = "true" ]; then
 			fi
 		done
 
-		for LOCKINGPID in `lsof | awk '{print $1" "$2}' | grep -E "/bin|/system|/data|/cache" | awk '{print $1}'`; do
-			BINARY=$(ps | grep " $LOCKINGPID " | grep -v "grep" | awk '{print $5}')
-			ECHOL "File ${BINARY} is locking a critical partition running as PID ${LOCKINGPID}, killing it now!"
-			EXECL kill -9 $LOCKINGPID
+		for LOCKINGPID in `/sbin/busybox lsof | awk '{print $1" "$2}' | grep -E "/bin|/system|/data|/cache" | awk '{print $1}'`; do
+			BINARY=$(cat /proc/${LOCKINGPID}/status | grep -i \"name\" | awk -F':\t' '{print $2}')
+			if [ "$BINARY" != "" ]; then
+				ECHOL "File ${BINARY} is locking a critical partition running as PID ${LOCKINGPID}, killing it now!" >> /tmp/xperiablfix.log
+				EXECL kill -9 $LOCKINGPID
+			fi
 		done
 
 		# umount partitions, stripping the ramdisk to bare metal
